frozen_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.10.0=py38h32f6830_0
albumentations=0.4.6=pypi_0
appdirs=1.4.4=py_0
argon2-cffi=20.1.0=py38h7b6447c_1
asciitree=0.3.3=py_2
astroid=2.4.2=py38_0
attrs=20.1.0=py_0
audioread=2.1.8=py38h32f6830_2
autopep8=1.5.4=pyh9f0ad1d_0
backcall=0.2.0=py_0
black=20.8b1=py_1
bleach=3.1.5=py_0
blinker=1.4=py38_0
bokeh=2.2.1=py38_0
brotlipy=0.7.0=py38h7b6447c_1000
bzip2=1.0.8=h7b6447c_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.6.20=hecda079_0
cachetools=4.1.1=py_0
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py38h32f6830_0
cffi=1.14.0=py38h2e261b9_0
chardet=3.0.4=py38_1003
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
cryptography=3.1=py38h1ba5d50_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py38_0
cytoolz=0.10.1=py38h7b6447c_0
dask=2.25.0=py_0
dask-core=2.25.0=py_0
dataclasses=0.7=py38_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
distributed=2.25.0=py38_0
efficientnet-pytorch=0.6.3=pypi_0
entrypoints=0.3=py38_0
expat=2.2.9=he6710b0_2
fasteners=0.15=py_0
ffmpeg=4.3=h167e202_0
flake8=3.8.3=py_1
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.2=h5ab3b9f_0
fsspec=0.8.0=py_0
future=0.18.2=py38h32f6830_1
geos=3.8.0=he6710b0_0
gettext=0.19.8.1=hd7bead4_3
glib=2.63.1=h5a9c865_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
google-auth=1.21.1=py_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.14=h23475e2_0
grpcio=1.31.0=py38hf8bcb03_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=hb1b8bf9_0
heapdict=1.0.1=py_0
icu=64.2=he1b5a44_1
idna=2.10=py_0
imageio=2.9.0=py_0
imgaug=0.4.0=py_1
importlib-metadata=1.7.0=py38_0
importlib_metadata=1.7.0=0
iniconfig=1.0.1=py_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=pyh9f0ad1d_1
isort=5.4.2=py38_0
jasper=1.900.1=hd497a04_4
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
jsonschema=3.2.0=py38_0
jupyter=1.0.0=py_2
jupyter_client=6.1.6=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py38_0
kiwisolver=1.2.0=py38hfd86e86_0
lame=3.100=h7b6447c_0
lazy-object-proxy=1.4.3=py38h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.34=h53a641e_7
libblas=3.8.0=17_openblas
libcblas=3.8.0=17_openblas
libclang=9.0.1=default_hde54327_0
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=hd88cf55_4
libflac=1.3.3=he1b5a44_0
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
liblapack=3.8.0=17_openblas
liblapacke=3.8.0=17_openblas
libllvm10=10.0.1=hbcb73fb_5
libllvm9=9.0.1=h4a3c616_1
libogg=1.3.2=h7b6447c_0
libopenblas=0.3.10=h5a2b251_0
libopencv=4.3.0=py38_2
libpng=1.6.37=hbc83047_0
libprotobuf=3.12.4=hd408876_0
librosa=0.8.0=pyh9f0ad1d_0
libsndfile=1.0.28=he1b5a44_1000
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libvorbis=1.3.6=h7b6447c_0
libwebp-base=1.1.0=h7b6447c_3
libxcb=1.14=h7b6447c_0
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.9=hea5a465_1
lightgbm=2.3.1=pypi_0
llvmlite=0.34.0=py38h269e1b5_4
locket=0.2.0=py38_1
lz4-c=1.9.2=he6710b0_1
mafat-radar-challenge=0.0.1=dev_0
markdown=3.2.2=py38_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h2af1d28_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
monotonic=1.5=py_0
more-itertools=8.5.0=py_0
msgpack-python=1.0.0=py38hfd86e86_1
mypy_extensions=0.4.3=py38_0
nbconvert=5.6.1=py38_0
nbformat=5.0.7=py_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
networkx=2.5=py_0
ninja=1.10.1=py38hfd86e86_0
nose=1.3.7=pypi_0
notebook=6.1.1=py38_0
nspr=4.22=he6710b0_0
nss=3.46.1=hab99668_0
numba=0.51.2=py38h0573a6f_1
numcodecs=0.6.4=py38he6710b0_0
numpy=1.19.1=py38hbc27379_2
oauthlib=3.1.0=py_0
olefile=0.46=py_0
opencv=4.3.0=py38_2
opencv-python-headless=4.4.0.42=pypi_0
openh264=2.1.1=h8b12597_0
openssl=1.1.1g=h516909a_1
packaging=20.4=py_0
pandas=1.1.2=py38h950e882_0
pandoc=2.10.1=0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
partd=1.1.0=py_0
pathspec=0.7.0=py_0
patsy=0.5.1=py38_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=7.2.0=py38hb39fc2d_0
pip=20.2.3=py_0
pixman=0.38.0=h7b6447c_0
pluggy=0.13.1=py38_0
pooch=1.1.1=py_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.7=py_0
prompt_toolkit=3.0.7=0
protobuf=3.12.4=py38he6710b0_0
psutil=5.7.2=py38h7b6447c_0
ptyprocess=0.6.0=py38_0
py=1.9.0=py_0
py-opencv=4.3.0=py38h23f93f0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pyflakes=2.2.0=py_0
pygments=2.6.1=py_0
pyjwt=1.7.1=py38_0
pylint=2.6.0=py38h32f6830_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.12.3=py38ha8c2ead_3
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.16.0=py38h7b6447c_0
pysocks=1.7.1=py38_0
pysoundfile=0.10.2=py_1001
pytest=6.0.2=py38h32f6830_0
python=3.8.5=h1103e12_7_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.4.0=py3.8_cuda10.0.130_cudnn7.6.3_0
pytorch-tabnet=1.2.0=pypi_0
pytz=2020.1=py_0
pywavelets=1.1.1=py38h7b6447c_0
pyyaml=5.3.1=py38h1e0a361_0
pyzmq=19.0.1=py38he6710b0_1
qt=5.12.5=hd8c4c69_1
qtconsole=4.7.7=pyh9f0ad1d_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
regex=2020.7.14=py38h7b6447c_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
resampy=0.2.2=py_0
resnest=0.0.6b20200915=pypi_0
rope=0.17.0=pyh9f0ad1d_0
rsa=4.6=py_0
scikit-image=0.16.2=py38h0573a6f_0
scikit-learn=0.23.2=py38hee58b96_0
scipy=1.5.2=py38h8c5af15_0
send2trash=1.5.0=py38_0
setuptools=49.6.0=py38_0
shapely=1.7.1=py38h98ec03d_0
six=1.15.0=py_0
sortedcontainers=2.2.2=py_0
sqlite=3.33.0=h62c20be_0
statsmodels=0.11.1=py38h7b6447c_0
tbb=2020.2=hfd86e86_0
tblib=1.7.0=py_0
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow-determinism=0.3.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=py38_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
timm=0.1.26=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
toolz=0.10.0=py_0
torch-summary=1.4.2=pypi_0
torchvision=0.5.0=py38_cu100
tornado=6.0.4=py38h7b6447c_1
tqdm=4.49.0=pyh9f0ad1d_0
traitlets=4.3.3=py38_0
tsfresh=0.17.0=py_0
typed-ast=1.4.1=py38h7b6447c_0
typing_extensions=3.7.4.3=py_0
urllib3=1.25.10=py_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py38h32f6830_1
wrapt=1.11.2=py38h7b6447c_0
x264=1!152.20180806=h7b6447c_0
xgboost=1.2.0=pypi_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zarr=2.4.0=py_0
zeromq=4.3.2=he6710b0_3
zict=2.0.0=py_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
